feat: add Pug template rendering support and configure views
This commit introduces Pug as the templating engine for the application. It includes the creation of a Pug template file (`index.pug`) and updates the `index.ts` file to configure the view engine and views directory. The changes enable rendering HTML using Pug templates instead of sending plain text responses.